requirements.txt
python-telegram-bot>=22.0
SQLAlchemy>=1.4
APScheduler>=3.9